Credits and Emailing!
Fixed the credits. Deleted Credits.cs, and replaced it with
CreditsScript.cs, which will increment the z movement on a GUIText
object in order to make credits "scroll up" the screen. I will edit the
credits' contents as needed. The content for the Credits' text is in
Assets\Resources\Fonts\Credits.txt inside this commit, and is pasted
manually into Unity.

Also added emailing functionality! The email function is in the Email.cs
. I sent 3 test emails to you, myself, and some friends, using test
data. The function call test is in Table.cs in this commit, in the Awake
function. The email function requires one of the players to have a GMAIL
account, since the code uses a gmail email domain to send the email.
Params are the Gmail account username for sender, his password, the
"To:" email addresses, the "From:" email address, the player's in-game
usernames, and the ending scores.

Please note that I had issues getting it emailing to work initially, due
to Google blocking the sign-in attempt. I had to goto
https://www.google.com/settings/security/lesssecureapps and disable the
"Secure Apps" feature to get the email to go through my Gmail account.

Also, you should make the score variable of the players a byte (saves
memory vs. an int), and add a "Username" variable for players in game. I
will also need to hide the email addresses in the "to/from" fields for
the final prototype. (SecureString datatype?)

Now working on Powerup handler and Announcer objects...

-Tamk1s